Fix mini_paint solution Clean our_mini_paint and our_micro_paint Add tests for mini_paint Fix return code in our_micro_paint and remove grep in tests

+++ mini_paint/mini_paint.c
@@ -53,7 +53,7 @@ int
 {
 	float	distance;
 
-	distance = sqrtf(powf(x - shape->x, 2.) + powf(y - shape->y, 2.));
+	distance = sqrtf((int)(powf(x - shape->x, 2.) + powf(y - shape->y, 2.)));
 	if (distance <= shape->radius)
 	{
 		if ((shape->radius - distance) < 1.00000000)
